# api/openapi.yaml
openapi: 3.0.3
info:
  title: Roadbook API
  description: >
    Read candidates and record decisions. This file is the contract (CLAUDE.md
    invariant 10): the Go server interface and the TypeScript client are both
    generated from it and never hand-written. Import and detection are CLI
    operations in phase 1 and do not appear here.
  version: 0.1.0
paths:
...
  /shared/{token}:
    get:
      operationId: getSharedAdventure
      summary: The read-only view a share link opens
      description: >
        The public read path (phase 13 BRIEF §3C): no session is consulted
        — the token is the credential, and this operation is exempt from
        the sign-in wall in every mode. It resolves the token to its
        decision, requires that decision to be confirmed and matched to a
        candidate of the owner's latest run, and answers with the same
        assembled journey the owner's own page draws: every leg carries its
        kind (invariants 5 and 8 apply to strangers most of all), countries,
        mode breakdown, and photos of both provenances with their
        placements. Unknown, revoked, and currently-unreachable links all
        answer the same 404 — a link that does not open tells nothing
        about why. The response carries X-Robots-Tag noindex; the page
        that renders it does too.
      parameters:
        - name: token
          in: path
          required: true
          description: The raw token from the link.
          schema:
            type: string
      responses:
        "200":
          description: The shared adventure.
          headers:
            X-Robots-Tag:
              schema: {type: string}
              description: Always "noindex, nofollow".
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedAdventure"
        "404":
          description: This link does not open anything.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /shared/{token}/photos/{id}/thumbnail:
    get:
      operationId: getSharedPhotoThumbnail
      summary: Thumbnail of an attached photo, by share token
      description: >
        Session-exempt like the shared view. The photo must belong to the
        adventure the token opens; any other photo id is 404 — the token
        grants exactly one adventure, never the photo id space.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: The thumbnail.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: The link does not open, or the photo is not on the shared adventure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /shared/{token}/import-photos/{id}/thumbnail:
    get:
      operationId: getSharedImportPhotoThumbnail
      summary: Thumbnail of a photo-import record, by share token
      description: >
        Session-exempt like the shared view. The record must be one the
        shared adventure lists (its capture inside the adventure's span,
        the owner's own record); anything else is 404, as is a record
        with no thumbnail (HEIC).
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: The thumbnail.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: The link does not open, the record is not on the shared adventure, or it has no thumbnail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/signin:
    get:
      operationId: startSignIn
      summary: Begin the OIDC sign-in redirect dance
      description: >
        Issues the provider redirect with a fresh state+nonce pair, carried
        in a short-lived HttpOnly cookie the callback verifies. A browser
        navigation target, not an API call — the generated client never
        invokes it.
      responses:
        "302":
          description: Redirect to the configured OIDC provider.
          headers:
            Location:
              schema: {type: string}
            Set-Cookie:
              description: The state cookie (10 minutes, HttpOnly).
              schema: {type: string}
        "409":
          description: This instance runs authless (mode off).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/callback:
    get:
      operationId: completeSignIn
      summary: OIDC provider redirect target
      description: >
        Exchanges the provider's code (server-to-server, with the client
        secret), verifies the ID token and nonce, upserts the user by
        subject, and mints an opaque session recorded in Postgres — the
        cookie carries the token, the table its hash.
      parameters:
        - name: code
          in: query
          schema: {type: string}
        - name: state
          in: query
          schema: {type: string}
        - name: error
          in: query
          description: Provider-reported failure (user denied, etc.).
          schema: {type: string}
      responses:
        "302":
          description: Signed in; back to the app.
          headers:
            Location:
              schema: {type: string}
            Set-Cookie:
              description: The session cookie (HttpOnly, SameSite=Lax).
              schema: {type: string}
        "400":
          description: State mismatch, missing code, or a token that does not verify.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: This instance runs authless (mode off).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
...
  schemas:
...
    LatLng:
      type: object
      required: [lat, lon]
      properties:
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
...
    TimedPoint:
      type: object
      required: [t, lat, lon]
      properties:
        t:
          type: string
          format: date-time
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
    Leg:
      type: object
      required: [kind, points, distance_km, start, end]
      properties:
        kind:
          type: string
          enum: [observed, gap]
        gap_kind:
          type: string
          enum: [unknown, road, air]
          description: >
            Present only when kind is "gap". "air": the gap's implied speed
            (endpoint chord over duration) met the air_speed_min_kmh
            parameter — rendered as a great-circle arc, never routed,
            excluded from road-distance validation. "road": the routing
            cache supplied road geometry for this gap. "unknown": neither —
            drawn as a dashed straight line that is not a route.
        points:
          type: array
          description: >
            An observed leg carries its full point run; a gap leg carries
            exactly its two endpoints.
          items:
            $ref: "#/components/schemas/TimedPoint"
        distance_km:
          type: number
          format: double
          description: Chord sum for observed legs; endpoint chord for gaps.
        routed_points:
          type: array
          description: >
            Road geometry from the routing cache — present only when
            gap_kind is "road". Rides alongside points (which still holds
            exactly the two timestamped endpoints): routed vertices have no
            timestamps, and are inference, not measurement.
          items:
            $ref: "#/components/schemas/LatLng"
        routed_km:
          type: number
          format: double
          description: >
            The routed road distance — present only when gap_kind is
            "road". distance_km keeps its chord meaning regardless.
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
    Stop:
      type: object
      required: [start, end, loc, points, displacement_km]
      properties:
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        loc:
          $ref: "#/components/schemas/LatLng"
        points:
          type: integer
        displacement_km:
          type: number
          format: double
          description: First-to-last straight line during the halt, not path sum.
    Country:
      type: object
      required: [iso_code, name]
      properties:
        iso_code:
          type: string
          description: >
            ISO 3166-1 alpha-2 where Natural Earth provides one (Kosovo's XK
            is user-assigned); territories without any alpha-2 carry Natural
            Earth's three-letter ADM0_A3 instead.
        name:
          type: string
    Journey:
      type: object
      required:
        [window_start, window_end, params, legs, stops, countries, total_km,
         observed_km, inferred_km, air_km, unknown_km, routed_km, ground_km,
         google_km, google_ground_km, merged_points, trace_points_kept,
         raw_points_kept]
      properties:
        window_start:
          type: string
          format: date-time
        window_end:
          type: string
          format: date-time
        params:
          type: object
          additionalProperties: true
          description: Exactly the parameters that produced this assembly (invariant 3).
        legs:
          type: array
          items:
            $ref: "#/components/schemas/Leg"
        stops:
          type: array
          items:
            $ref: "#/components/schemas/Stop"
        countries:
          type: array
          description: >
            Countries the route's points fall in, ordered by first
            appearance along the journey. Derived locally by point-in-polygon
            against bundled Natural Earth polygons — never a network lookup.
            Empty when `roadbook countries` has not been run. Border-adjacent
            points can misattribute at 1:110m resolution; the UI labels this
            line as derived.
          items:
            $ref: "#/components/schemas/Country"
        total_km:
          type: number
          format: double
        observed_km:
          type: number
          format: double
        inferred_km:
          type: number
          format: double
        air_km:
          type: number
          format: double
          description: >
            Chord sum over air-classified gaps — a subset of inferred_km.
            The endpoint chord is the great-circle distance, so it is also
            the arc's length. Excluded from road-distance validation. 0 for
            a ground journey.
        unknown_km:
          type: number
          format: double
          description: >
            Chord sum over gaps still unknown after the routing cache was
            consulted — the visibly-unfilled remainder. A router that cannot
            fill a gap leaves it here by design (patchy OSM coverage is the
            expected case, not an error).
        routed_km:
          type: number
          format: double
          description: >
            Routed road distance summed over road legs. 0 when nothing is
            routed — the product is fully usable without a router.
        ground_km:
          type: number
          format: double
          description: >
            The road-comparable reconstruction: observed + routed + unknown
            chords. Air legs are excluded by construction.
        google_ground_km:
          type: number
          format: double
          description: >
            Google's own activity distances minus FLYING-mode activities —
            the comparable side of the road validation. Air must leave both
            sides or neither. Mode is a guess: a flight Google mislabelled
            as ground inflates this figure, and the divergence flag is the
            tripwire that surfaces it.
        mode_breakdown:
          type: array
          description: >
            Source-asserted per-mode distance (phase 11 §6.2): the window's
            activity distances summed by Google's own mode labels, ordered
            km-descending. These are the source's guesses (recorded
            failures: a 1,023 km "motorcycling" relocation, probability-0.00
            modes; the pipeline itself trusts speed over mode for flight),
            so display keeps them visually subordinate, labels the source,
            and never sums them against measured geometry. Absent — not
            empty — when the journey's evidence carries no activities at
            all (photo-sourced journeys): the display states that rather
            than showing zeros.
          items:
            $ref: "#/components/schemas/ModeKm"
        divergence_pct:
          type: number
          format: double
          description: >
            Signed percentage by which ground_km differs from
            google_ground_km. Absent when Google recorded no ground
            distance in the window — nothing to compare is not zero
            divergence. A conversation starter, never a gate: unroutable
            gaps under-count, OSM detours over-count, and both explain
            themselves on the map.
        divergence_flagged:
          type: boolean
          description: >
            True when divergence_pct exceeds the divergence_warn_pct
            parameter (echoed in params) in either direction. Absent
            whenever divergence_pct is absent. Computed server-side so the
            frontend renders the flag without owning the threshold.
        google_km:
          type: number
          format: double
          description: >
            Google's own distance figure summed over the window's activities —
            the independent check routed distance is validated against in
            phase 3. 0 when the window holds no activities.
        merged_points:
          type: integer
        trace_points_kept:
          type: integer
        raw_points_kept:
          type: integer
    Photo:
      type: object
      required:
        [id, original_name, time_source, pos_source, thumb_w, thumb_h,
         uploaded_at]
      properties:
        id:
          type: integer
          format: int64
        original_name:
          type: string
        taken_at:
          type: string
          format: date-time
          description: >
            The resolved capture instant. Absent when no rung of the time
            ladder resolved (time_source "none") — the photo is stored and
            shown, but unplaced on map and timeline.
        taken_offset_sec:
          type: integer
          description: >
            Civil offset for display, seconds east of UTC. Present exactly
            when taken_at is.
        time_source:
          type: string
          enum: [gps, exif_offset, sidecar, exif_local, none]
          description: >
            Which rung of the resolution ladder produced taken_at (BRIEF
            §3E), strongest first — the GPS receiver's own UTC clock, the
            explicit EXIF offset tag, the Takeout sidecar's epoch, or the
            wall clock interpreted in the adventure's own offset (stated as
            the weakest rung; wrong by the zone difference on multi-zone
            adventures).
        pos:
          $ref: "#/components/schemas/LatLng"
        pos_source:
          type: string
          enum: [exif, sidecar, none]
          description: >
            Which reading produced pos (BRIEF §3D) — the camera's own GPS
            block, or Google's copy from the Takeout sidecar. "none": no
            position; the photo appears in the strip but not on the map.
        thumb_w:
          type: integer
          description: Thumbnail pixel width, post-orientation.
        thumb_h:
          type: integer
        uploaded_at:
          type: string
          format: date-time
        place_kind:
          type: string
          enum: [observed, road, unknown, air, stop]
          description: >
            Which element of the journey held this photo's instant — and
            therefore which drawn geometry distance_from_route_m was
            measured against (BRIEF §3G) - an observed leg's measured
            points, a road gap's routed polyline, an unknown gap's chord, or
            a stop's location. Absent when the photo is unplaced: no
            position, no resolved instant, or an instant outside every leg
            and stop. Derived at read time, never stored — re-detection
            re-places photos automatically.
        leg_index:
          type: integer
          description: Index into the journey's legs. Present for leg place kinds.
        stop_index:
          type: integer
          description: Index into the journey's stops. Present when place_kind is "stop".
        distance_from_route_m:
          type: number
          format: double
          description: >
            Point-to-polyline distance from the photo's position to the
            placed element's drawn geometry. Absent for place_kind "air":
            the arc is presentation between two endpoints, not a claimed
            path, so "far from it" asserts nothing.
        far_flagged:
          type: boolean
          description: >
            True when distance_from_route_m exceeds the photo_far_warn_m
            parameter (echoed in the list's params). The photo's position
            is a measurement — the most accurate in the project — so the
            flag marks a disagreement between the photo and the inferred
            route, not a doubt about the photo. A conversation starter,
            never a gate.
...
    ModeKm:
      type: object
      required: [mode, km]
      properties:
        mode:
          type: string
          description: The source's own label (IN_BUS, FLYING, …), unreworded.
        km:
          type: number
          format: double
    ImportPhoto:
      type: object
      required:
        [id, import_id, original_name, taken_at, taken_offset_sec,
         time_source, pos, pos_source, thumb_w, thumb_h, uploaded_at]
      description: >
        One photo-import record (phase 11 §4D), distinct from Photo
        deliberately: records ride imports and are never deletable from an
        adventure, while attached photos ride decisions and are — one schema
        for both would blur what each id can do. Records exist only for
        usable photos, so position and instant are always present.
      properties:
        id:
          type: integer
          format: int64
        import_id:
          type: integer
          format: int64
          description: The photo import this record arrived in.
        original_name:
          type: string
        taken_at:
          type: string
          format: date-time
        taken_offset_sec:
          type: integer
          description: Civil offset for display, seconds east of UTC.
        time_source:
          type: string
          enum: [gps, exif_offset, sidecar, exif_local]
        pos:
          $ref: "#/components/schemas/LatLng"
        pos_source:
          type: string
          enum: [exif, sidecar]
        thumb_w:
          type: integer
          description: Thumbnail pixel width; 0 means no thumbnail (HEIC).
        thumb_h:
          type: integer
        uploaded_at:
          type: string
          format: date-time
        place_kind:
          type: string
          enum: [observed, road, unknown, air, stop]
          description: >
            Same read-time placement as Photo.place_kind, against the same
            drawn geometry. Absent when the instant falls outside every leg
            and stop of the assembled journey.
        leg_index:
          type: integer
        stop_index:
          type: integer
        distance_from_route_m:
          type: number
          format: double
        far_flagged:
          type: boolean
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SharedAdventure:
      type: object
      description: >
        Everything the shared plate renders, and nothing about the owner:
        no candidate id, no score, no account. The photo entries' ids are
        meaningful only through the token-scoped thumbnail operations.
      required:
        [name, span_start, span_end, start_truncated, end_truncated, journey,
         photos, import_photos, params]
      properties:
        name:
          type: string
          description: The adventure's confirmed name.
        span_start:
          type: string
          format: date-time
        span_end:
          type: string
          format: date-time
        start_truncated:
          type: boolean
        end_truncated:
          type: boolean
        journey:
          $ref: "#/components/schemas/Journey"
        photos:
          type: array
          description: Photos attached to the adventure, with placements.
          items:
            $ref: "#/components/schemas/Photo"
        import_photos:
          type: array
          description: Photo-import records span-joined to the adventure, with placements.
          items:
            $ref: "#/components/schemas/ImportPhoto"
        params:
          type: object
          additionalProperties: true
          description: The photo placement parameters (invariant 3) — photo_far_warn_m.